Moved for-credit group membership tracking to the group_memberships table.
This completely eliminates the need for the CreditGroup subclass.
Rather than for-credit being tracked as its own group of people,
"credit" is a column in the group_memberships table. This is a much cleaner solution,
as a "for credit" dedicated group will always be a subset of another group.
There was also some leftover code in the Student class from when "credit" was tracked
by student which has been cleaned up.

Signed-off-by: Joe Baker <jbaker0428@gmail.com>