Parsing timestamp strings with the python-dateutils module.
The convert_timestamp method is not part of APSW, only the default sqlite module.
Even if the sqlite module were used, convert_timestamp does not handle timestamps properly.

This also makes it easier to deal with timezone differences -- parts of the gcal data are UTC.

NOTE: The RFID reader outputs data in the music cabinet laptop's local time zone without tzinfo output.
Our code currently assumes that to always be EST... what if the music cabinet was configured to UTC?

Signed-off-by: Joe Baker <jbaker@alum.wpi.edu>